refactor: Move AccessEgresses to street package
